feat: AbilityCostBase no longer Activates on ApplyCost trigger (see notes) - Network settings are now available again
notes: the Ability is already listening for the cost trigger OnGive.
Moreover, clients will Activate the Cost Ability with no means of
Deactivating it. The Ability now also Applies the cost only on Hosts.
Since the network settings were never available, AbilityCost was almost
always applicable on the host only (unless playing SinglePlayer).
Further changes will allow it to be evaluated on Clients based on the
ReplicationType and ExecutionPolicy