fix: When handling an ability transfer, process the removal before ensuring that the recipient does not have the ability
This change is to adress the edge case where a bullet contained within a magazine that is contained in an inventory is moved out of the magazine and into the inventory directly. If the magazine is extending the inventory, then the ensure would fire since the inventory technically already has the bullet ability before the removal.