fix: When running Min/Max processor, add a PreviousValue fragment for the "Current" of the given attribute IFF it does not exist yet.
This fragment is normally added by the StorePreviousValue processor. However, that processor operates on the attribute component (min/max/current) if it was asked to recompute.

In the case where the current as not been changed yet, but other attribute component have, then the required PreviousValue fragment for the current will not exist yet. In such a case the Min/Max Clamp processor will create one with the current value of the Current fragment