fix: Signal payload is now stored before calling broadcast instead of after
This fixes a fringe edge case where, in response to the signal being broadcasted, someone would immediately bind to the same signal. Because this would happen in the same "stack", since there is no payload being stored yet, if the bind was done with the policy "FirePayloadInFlight" or "FirePayloadInFlightThisFrame", the delegate being passed would never be invoked.