fix: Ensures now break into the debugger with the correct callstack
notes: we no longer use the callback function from the custom window to
break and instead wait on the result of the button presses to determine
if 'Break' was selected. After which, we break as normal. This makes
sure that the custom window callstack is not part of the
break-in-debugger