fix: TransferExistingAbility now makes a Request from the Ability to process the Transfer
notes: read notes in code on why we need to do the OnershipTransfer and
set the Script's ContextEntityWithActor variable immediately and BEFORE
the Request is resolved in the Ability processor

We also issue a transfer request for ALL the sub-abilities (if any) of
the Ability we are transferring. This ensures that the sub-abilities go
through their Deactivate and Revoke cycle and Given again. Without this,
sub-abilities such as Conditions will not re-evaluate when the Transfer
is first completed