feat: Memory Stats are now running on another thread and should no longer stall the game thread
notes: the Memory Stats can be quite expensive for a large level and can
cause stalls. We could turn off the stats, however they are quite useful
during development. Instead, we now perform the stats gathering in
another thread.

The downside to the current implementation (not a big one) is that the
thread runs all the time, even when the game is not running.