fix: Abilities now ignore Activate/Deactivate Requests after Revoke
notes: this fixes issues where there could be pending Activate events in
the same frame as a Revoke resulting in ensures firing when trying to
Activate an Ability that has already been Revoked. Ideally, we would not
add the Request in the first place, but perhaps this is still a good
approach especially if this needs to be debugged in the future.