feat: removed SyncFromActor tag
notes: this tag, at the moment, is no longer useful since we now have
SceneNodes which need updating every frame. The tag was also used as a
hidden optimization but now that our Transform is playing a more central
role for our Entities, the hidden optimization does not make sense since
we need to update the transform every frame.

Discussion points:
- The SyncToActor/SyncFromActor being hidden is not great (we are being
clever with the SyncToActor tag right now)
- In DnD DA the Transform processor ended up being costly, let's keep
that in mind
- If we expose the setting (like in Mass), this basic feature becomes a
lot hard to use
- For now, we are always going to sync and revisit + re-introduce this
functionality later as it is slowing us down when it comes to completing
SceneNode and other tasks. Hopefully we come up with a stronger design
and implementation that is performance friendly