fix: removed the hack that no longer appears to be necessary
notes: this was a hack to work around issues with dragging Pins from
details panel to the node and drag/drop connecting. It's possible that
the derived K2Nodes had some erroenous code (see upcoming commit of
EntityScript K2Node) which forced this hack to be present. We'll
continue investigating this if any issues with our custom K2Nodes
surface again.