feat: Base Ck_UserWidget now implements the ICk_Entity_ContextInjector_Interface interface
The current Ck_UserWidget is setup to expect the context to be a BindActor first.

Moving forward, this should be removed in favor of a bind/context entity as it is more flexible and works better with the rest of the framework and feature set.

We also don't want to require users to "have to" use our base widget class. This class should be treated as an accelerant only.