exp: tried to have the Ism Renderer created automatically from Proxy
notes: the main issue is that there are too many options unique to the
IsmComponent of Unreal that will need to be either dropped or somehow
each Proxy will have to define them (so that we can pick the correct
Renderer for it).

We _could_ get rid of some of the options e.g. Physics. But, other
options, such as Material, Lighting and Culling will need to be defined
by the Proxy. Maybe that's okay and each combination of settings creates
a unique IsmProxy.