fix: EntityScript K2Node now correctly serializes the InLifetimeOwner pin and subsequent connections when loading a saved BP
notes: the main issue was that we were ignoring the existing state,
which is passed to us by the `InOldPins`. The next issue is that the
dynamic pin is not constructed when constructing default pins. If a pin
is not constructed when constructing default pins, it is marked as an
'bOrphan' and subsequently is unable to link up correctly when compiling
the node.

We did try a hack (which worked) which was to force set `bOrphan` to `0`
which actually did fix the issue. However, it felt like a hack that
could be problematic later on. Most likely, we are missing a setting (we
did try `OrphanedPinSaveMode` to `SaveAll`, however that did not work)
that we do not yet know about which allows orphaned pins to also
serialize correctly and then not be marked as orphaned when
de-serializing and loading the graph and the node.

The current solution is to treat it as a default pin. And we can do this
now because we are using the old state of the node to figure out whether
we need the dynamic pin or not.