feat: added ReplicationData_EntityScript replicated variable with RepNotify
notes: the new function follows a similar pattern to our existing
replicate from ConstructionObject. Difference being that we make a
request to the EntityScript processor to complete the spawn.

There is an unresolved issue with this approach and that's the number of
dependents. When we defer to the processor, the number of dependents
does not tally up. We'll need a clean solution for this issue.