feat, fix: EntityScript now has a PostConstruction function option which is used by the ReplicationDriver - ReplicationDrive is no longer leaking the AssociatedEntity
notes: ReplicationDriver always has an Entity assigned before the
RepNotifies are invoked. Previously, we were discarding that Entity
(without destroying it) and replacing it with a new Entity.

The EntityScript Utils now has an overloaded function that takes in an
existing Entity and 'Spawns' an EntityScript on the existing Entity.