fix: Handle DeQuantize now checks whether the ReplicatedObject is linked before proceeding
notes: if the ReplicatedObject is not linked, this means that the Handle
is being replicated earlier than expected. With ReplicatedVariables,
this should not be a problem but if we are RPCing and the Client isn't
ready (as an example) then this path will trigger and the Handle will
not replicate correctly.

Example: this fix arose from an issue where replicated projectiles were
being fired immediately upon the starting the level and some of the
projectiles were hitting a wall, getting destroyed AND firing a cue with
the Instigator Handle (Instigator being the projectile spawner). All of
this happened while the Client is still getting ready. Consequently, the
RPC for the cue, with the Instigator as one of the replicated variables,
fires too early resulting in an ensure firing (before these changes).
Now, we return early and do not attempt to replicate the Handle if the
replicated objects have not been linked.