fix: simplified the EntityScript construction and fixed replication of 'number of dependents' count
notes: the EntityScript construction was divided up when CkNet existed
as a separate submodule. Now that we have consolidate CkNet into CkEcs,
we can simplify some of the flow and eliminate some bugs. One such bug
is the replication of the number of dependents which was not being
correctly done because the Replication was being done before
EntityScript Construct was being called.