fix: ReplicationComplete variants no longer return true if the ReplicationDriver is not found
notes: the initial thought process was that in SinglePlayer, we will
NOT have a ReplicationDriver and thus this should return true. However,
it's also possible to not have the ReplicationDriver added in time while
the Entity is being constructed for _certain_ (rare) cases. With that in
mind, it makes sense to not have this be relied on as a bypass for
Singleplayer. Instead, we now assume that if this is called, the
ReplicationDriver exists.

This of course breaks some existing code, which now needs to perform
additional checks which we think is the right way to go to make the code
more clear and less error prone.