fix: fixed issue where EntityCollection replication might be rejected due to faulty if-logic
notes: while in OnRep, the Entities should only be checked for the
ReplicationDriver, which MUST be present for the Entity to successfully
replicate. While in the processor, the Entities should be checked for
replication completeness, so that when the EntityCollection callbacks
are invoked, that the Entity is guaranteed to be ReplicationComplete.