fix: ResourceLoader auto-loading of assets is no longer Async
notes: the Async loading causes some assets, particularly
BlueprintNodeTemplate assets, to not be loaded in time for the
ClassActions to be refreshed (i.e. the Task Node to appear in the right
click Actions menu). Surprisingly, forcing the RefreshClassActions when
the assets are loaded (e.g. in the callback on BatchLoading in the
ResourceLoader Processor) does not fix the problem of some (in our case,
most) TaskNodes not appearing in the right-click Actions menu.

Note that we are already compensating for an issue with the Task Nodes
where if they are not loaded they do not show up in the Actions menu.
One of the reasons we have the ResourceLoader Subsystem.