hack-fix: hack fix for an issue with moving Registry from one Handle to another
notes: when creating payloads for Signals, it is possible to send the
same Handle to the Signal _and_ the Payload where the Payload forwards
the handle which eventually gets moved. Unfortunately, this move results
in the original Handle (up the call stack) to lose it's Registry
pointer. This points to a problem with the move semantics either in our
FCk_Registry OR the internal registry pointer.

Either way, we don't have the bandwidth to fix this now. We'll take a
look at this problem later. Until then, we'll pay the cost of copying
instead of moving.