build-fix: cook fix for EntityScript generated structs
notes: saving the structs while cooking results in a crash in
AsyncLoading where the object seems to not be fully parsed and the
ExportMap not fully populated resulting in a hard-crash of the cooker.

Other fixes include not running the OnAssetAdded code if cooking. We
also force load all Params assets in the Params folder for the
EntityScript per module.