feat: added support for debugging using Jolt's debugger directly
notes: this is useful in contexts where the Probes are not working as
expected and we want to double check that our world and Jolt world are
seeing things exactly the same.

The Jolt debugger is VERY slow (rather our implementation of it) and
should not be used normally over our regular Probe debugger.